--- src/components/UserRegister/index.tsx
@@ -1,7 +1,0 @@
-import UserRegisterForm from './UserRegisterForm';
-
-const UserRegister = () => {
-  return <UserRegisterForm />;
-};
-
-export default UserRegister;
