+++ src/components/UserRegisterForm/Fields/IsMember.tsx
@@ -2,9 +2,10 @@ import { css } from '@emotion/react';
 import { isBoolean } from 'is-what';
 
 import { Button, SsafyIcon } from '~/components/Common';
-import Question from '~/components/UserRegister/Question';
 import { useUpdateMyInfoFormContext } from '~/services/member';
 import { flex } from '~/styles/utils';
+
+import Question from '../Question';
 
 const fieldName = 'ssafyMember';
