+++ src/components/UserRegisterForm/Fields/Year.tsx
@@ -2,9 +2,10 @@ import { css } from '@emotion/react';
 import { useEffect } from 'react';
 
 import { SelectBox } from '~/components/Common';
-import Question from '~/components/UserRegister/Question';
 import { useUpdateMyInfoFormContext } from '~/services/member';
 import { flex } from '~/styles/utils';
+
+import Question from '../Question';
 
 const years = Array(10)
   .fill(undefined)
