+++ src/pages/auth/register/index.tsx
@@ -4,7 +4,7 @@ import { useRouter } from 'next/router';
 
 import { DefaultFullPageLoader } from '~/components/Common';
 import UserRegisterLayout from '~/components/Layout/UserRegisterLayout';
-import UserRegister from '~/components/UserRegister';
+import UserRegisterForm from '~/components/UserRegisterForm';
 import { useMyAccountStatus } from '~/services/member';
 
 const RegisterPage: CustomNextPage = () => {
@@ -17,7 +17,7 @@ const RegisterPage: CustomNextPage = () => {
 
   return (
     <UserRegisterLayout>
-      <UserRegister />
+      <UserRegisterForm />
     </UserRegisterLayout>
   );
 };
